+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * **Breaking change**: Configuration options `filter.flate_compression` and
   `filter.flate_memory` are changed to `filter.flate.compression` and
   `filter.flate.memory`
+* **Breaking change**: [HexaPDF::Document#wrap] handles cross-reference and
+  object stream specially to avoid problems with invalid PDFs
 * [HexaPDF::Composer::new] to allow skipping the initial page creation
 * CLI command `hexapdf info --check` to process streams to reveal stream errors